backend/api/cv_model.py
import cv2
import numpy as np
# import tensorflow as tf  # Commented out for now
from django.conf import settings
import os

class TrashDetectionModel:
    def __init__(self):
        self.model = None
        self.class_names = [
            'bottle', 'can', 'paper', 'plastic_bag', 'cigarette', 
            'food_waste', 'glass', 'metal', 'cardboard', 'other_trash'
        ]
        self.colors = {
            'bottle': '#FF6B6B',
            'can': '#4ECDC4', 
            'paper': '#45B7D1',
            'plastic_bag': '#96CEB4',
            'cigarette': '#FECA57',
            'food_waste': '#FF9FF3',
            'glass': '#54A0FF',
            'metal': '#5F27CD',
            'cardboard': '#00D2D3',
            'other_trash': '#FF3838'
        }
        self.load_model()
    
    def load_model(self):
        """
        Load the trash detection model
        For now, we'll use a mock model. In production, you would load your trained model here.
        """
        try:
            # Mock model for demonstration
            # Replace this with your actual model loading code
            # self.model = tf.keras.models.load_model('path/to/your/model.h5')
            print("Mock trash detection model loaded")
        except Exception as e:
            print(f"Error loading model: {e}")
    
    def preprocess_image(self, image_path):
        """Preprocess image for model input"""
        try:
            image = cv2.imread(image_path)
            image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
            image = cv2.resize(image, (416, 416))  # Adjust size based on your model
            image = image.astype(np.float32) / 255.0
            image = np.expand_dims(image, axis=0)
            return image
        except Exception as e:
            print(f"Error preprocessing image: {e}")
            return None
    
    def detect_trash(self, image_path):
        """
        Detect trash in the given image
        Returns: list of detected objects with bounding boxes and confidence scores
        """
        try:
            # Mock detection results for demonstration
            # Replace this with actual model inference
            mock_detections = [
                {
                    'class': 'bottle',
                    'confidence': 0.85,
                    'bbox': [100, 150, 200, 250],  # [x1, y1, x2, y2]
                    'color': self.colors['bottle']
                },
                {
                    'class': 'plastic_bag',
                    'confidence': 0.72,
                    'bbox': [300, 100, 400, 180],
                    'color': self.colors['plastic_bag']
                }
            ]
            
            # Actual model inference would look like:
            # preprocessed_image = self.preprocess_image(image_path)
            # if preprocessed_image is not None:
            #     predictions = self.model.predict(preprocessed_image)
            #     detections = self.postprocess_predictions(predictions)
            #     return detections
            
            return mock_detections
            
        except Exception as e:
            print(f"Error in trash detection: {e}")
            return []
    
    def postprocess_predictions(self, predictions):
        """Process model predictions into readable format"""
        # This would contain your actual postprocessing logic
        # based on your model's output format
        detections = []
        # Process predictions...
        return detections

# Global model instance
trash_detector = TrashDetectionModel()
