backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pymongo==4.6.0
django-cors-headers==4.3.1
Pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
python-dotenv==1.0.0
